+++ src/test/resources/it/pagopa/pn/cucumber/workflowNotifica/b2b/giacenzaAttoGiudiziario890/GestioneGiacenzaAttoGiudiziario890.feature
@@ -369,21 +369,6 @@ Feature: avanzamento notifiche b2b con workflow cartaceo gestione giacenza atto 
     Then vengono letti gli eventi fino all'elemento di timeline della notifica "SEND_ANALOG_PROGRESS" con deliveryDetailCode "RECAG010"
 
   @giacenza890Simplified
-  Scenario: [MATTEO]
-    Given viene generata una nuova notifica
-      | subject            | notifica analogica con cucumber |
-      | senderDenomination | Comune di palermo               |
-    And destinatario
-      | denomination            | OK-Giacenza-gt10_890 |
-      | taxId                   | CLMCST42R12D969Z     |
-      | digitalDomicile         | NULL                 |
-      | physicalAddress_address | Via@ok_AR            |
-    When la notifica viene inviata tramite api b2b dal "Comune_Multi" e si attende che lo stato diventi ACCEPTED
-    Then vengono letti gli eventi fino all'elemento di timeline della notifica "SEND_ANALOG_PROGRESS" con deliveryDetailCode "CON018"
-    And Lato utente l'elemento di timeline della notifica "SEND_ANALOG_PROGRESS" con deliveryDetailCode "CON018" non è visibile
-
-
-  @giacenza890Simplified
   Scenario: [B2B_GIACENZA_890_WI2.2_17] Invio notifica con sequence @OK-Giacenza_RS ed attesa elemento di timeline SEND_ANALOG_PROGRESS con deliveryDetailCode RECRS010
     Given viene generata una nuova notifica
       | subject            | notifica analogica con cucumber |
